Crate blue_pill

Board Support Crate for the Blue Pill

Usage

Follow cortex-m-quickstart instructions but remove the memory.x linker script and the build.rs build script file as part of the configuration of the quickstart crate.

Reexports

pub extern crate stm32f103xx;

pub use capture::Capture;

pub use pwm::Pwm;

pub use qei::Qei;

pub use rtc::Rtc;

pub use serial::Serial;

pub use spi::Spi;

pub use timer::Channel;

pub use timer::Timer;

Modules

adc
Analog to Digital Converter

ahb
Advance High-performance Bus (AHB)

apb1
Advance Peripheral Bus 1 (APB1)

apb2
Advance Peripheral Bus 2 (APB2)

capture
Input capture interface

dma
Direct Memory Access (DMA)

gpio
General Purpose I/O

led
User LEDs

prelude
The prelude is a collection of all the traits in this crate

pwm
Pulse Width Modulation

qei
Quadrature Encoder Interface (QEI)

rtc
Real Time Clock

serial
Serial interface

spi
Serial Peripheral Interface

time
Units of time

timer
Timer