Struct blue_pill::adc::Adc1

pub struct Adc1<'a>(pub &'a ADC1);

ADC Channel 1 (PA1)

Methods

impl<'a> Adc1<'a>

pub fn init(&self, dma1: &DMA1, gpioa: &GPIOA, rcc: &RCC)

Initializes the ADC

NOTE Pwm<TIM2>.init must be called before this method because both methods configure the PA1 pin (one as input and the other as output :-/)

pub fn disable(&self)

Disables the ADC

pub fn enable(&self)

Enables the ADC

pub fn start<B>(
&self,
buffer: &Static<CircBuffer<B, Dma1Channel1>>,
dma1: &DMA1,
pwm: Pwm<TIM2>
) -> Result<(), Error> where
B: Unsize<[u16]>,

Starts an analog to digital conversion that will be periodically triggered by the channel 2 of TIM2

The conversions will be stored in the circular buffer