Module blue_pill::dma

Direct Memory Access (DMA)

Structs

Buffer
Buffer to be used with a certain DMA CHANNEL

CircBuffer
A circular buffer associated to a DMA CHANNEL

Dma1Channel1
Channel 1 of DMA1

Dma1Channel2
Channel 2 of DMA1

Dma1Channel4
Channel 4 of DMA1

Dma1Channel5
Channel 5 of DMA1

Ref
Wraps a borrowed reference to a value in a Buffer

RefMut
A wrapper type for a mutably borrowed value from a Buffer`

Enums

Error
DMA error