Struct blue_pill::timer::Timer

pub struct Timer<'a, T>(pub &'a T)
 where
    T: 'a;

hal::Timer implementation

Methods

impl<'a> Timer<'a, TIM1>

pub fn init<P>(&self, period: P, rcc: &RCC) where
P: Into<Ticks>,

Initializes the timer with a periodic timeout of frequency Hz

NOTE After initialization, the timer will be in the paused state.

impl<'a, T> Timer<'a, T> where
T: Any + TIM,

pub fn init<P>(&self, period: P, rcc: &RCC) where
P: Into<Ticks>,

Initializes the timer with a periodic timeout of frequency Hz

NOTE After initialization, the timer will be in the paused state.